requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langgraph
pandas
openpyxl
python-dotenv
pydantic